Migrate web/Storybook builds from Rspack to Rsbuild
Rsbuild wraps Rspack with higher-level config (source.define, output.copy,
server.https/proxy, performance.buildCache) and is already used for Storybook,
so unifying onto it removes the last hand-rolled dev-server/HTTPS/persistent-cache
plumbing in config/rspack/ and lets the app build and Storybook share one
getSharedConfiguration()/getCommonConfiguration() config instead of two
independently-maintained bundler configs.
