+++ .gitignore
@@ -129,7 +129,7 @@ dist/
 web-build/
 
 # Local https certificate/key
-config/rspack/*.pem
+config/rsbuild/*.pem
 
 # Rspack build profiling output (RSPACK_PROFILE=OVERVIEW|ALL)
 .rspack-profile-*/
