+++ config/rsbuild/CustomVersionFilePlugin.ts
