+++ config/rsbuild/ForceGarbageCollectionPlugin.ts
