+++ config/rsbuild/ModuleInitTimingPlugin.ts
