--- config/rspack/rspack.common.ts
@@ -1,497 +1,0 @@
-import type {Configuration, DefinePluginOptions, RspackPluginInstance, SwcJsMinimizerRspackPluginOptions} from '@rspack/core';
-
-import {GenerateSW} from '@aaroon/workbox-rspack-plugin';
-import {rspack} from '@rspack/core';
-import {sentryWebpackPlugin} from '@sentry/webpack-plugin';
-import {execSync} from 'child_process';
-import dotenv from 'dotenv';
-import fs from 'fs';
-import {createRequire} from 'module';
-import path from 'path';
-import {fileURLToPath} from 'url';
-import {BundleAnalyzerPlugin} from 'webpack-bundle-analyzer';
-
-import type Environment from './types.ts';
-
-// @ts-expect-error -- Can't use .ts extensions without allowImportingTsExtensions in tsconfig
-import CustomVersionFilePlugin from './CustomVersionFilePlugin.ts';
-// @ts-expect-error -- Can't use .ts extensions without allowImportingTsExtensions in tsconfig
-import ModuleInitTimingPlugin from './ModuleInitTimingPlugin.ts';
-// @ts-expect-error -- Can't use .ts extensions without allowImportingTsExtensions in tsconfig
-import RspackPreloadPlugin from './RspackPreloadPlugin.ts';
-
-const require = createRequire(import.meta.url);
-const filename = fileURLToPath(import.meta.url);
-const dirname = path.dirname(filename);
-
-dotenv.config();
-
-function getCurrentBranchName(): string {
-    try {
-        return execSync('git rev-parse --abbrev-ref HEAD', {encoding: 'utf-8'}).trim();
-    } catch {
-        return '';
-    }
-}
-
-const localBranchName = getCurrentBranchName();
-
-const includeModules = [
-    'react-native-reanimated',
-    'react-native-worklets',
-    'react-native-picker-select',
-    'react-native-web',
-    'react-native-webview',
-    '@react-native-picker',
-    '@react-navigation/material-top-tabs',
-    '@react-navigation/native',
-    '@react-navigation/native-stack',
-    '@react-navigation/stack',
-    'react-native-gesture-handler',
-    'react-native-google-places-autocomplete',
-    'react-native-qrcode-svg',
-    'react-native-view-shot',
-    '@react-native/assets',
-    'expo',
-    'expo-audio',
-    'expo-video',
-    'expo-image',
-    'expo-image-manipulator',
-    'expo-modules-core',
-    'victory-native',
-    '@shopify/react-native-skia',
-].join('|');
-
-const environmentToLogoSuffixMap: Record<string, string> = {
-    production: '-dark',
-    staging: '-stg',
-    dev: '-dev',
-    adhoc: '-adhoc',
-};
-
-function mapEnvironmentToLogoSuffix(environmentFile: string): string {
-    let environment = environmentFile.split('.').at(2);
-    if (typeof environment === 'undefined') {
-        environment = 'dev';
-    }
-    return environmentToLogoSuffixMap[environment];
-}
-
-/**
- * `DefinePlugin`/Rsbuild `source.define` values shared between the main app build (via
- * `rspack.DefinePlugin` below) and Storybook (via `.storybook/rsbuild.config.ts`'s `source.define`),
- * so both bundle the same `__DEV__`/`__REACT_WEB_CONFIG__`/`__GIT_BRANCH__` values for a given env file.
- */
-function getDefineValues(file: string): DefinePluginOptions {
-    const isDevelopmentFile = file === '.env' || file === '.env.development';
-    /* eslint-disable @typescript-eslint/naming-convention */
-    return {
-        process: {env: {}},
-        // Define EXPO_OS for web platform to fix expo-modules-core warning
-        'process.env.EXPO_OS': JSON.stringify('web'),
-        __REACT_WEB_CONFIG__: JSON.stringify(dotenv.config({path: file}).parsed),
-
-        // React Native JavaScript environment requires the global __DEV__ variable to be accessible.
-        // react-native-render-html uses variable to log exclusively during development.
-        // See https://reactnative.dev/docs/javascript-environment
-        __DEV__: /staging|prod|adhoc/.test(file) === false,
-        // Expose the current git branch so the debug menu can display it in the browser tab title.
-        // Empty string in non-development builds.
-        __GIT_BRANCH__: JSON.stringify(isDevelopmentFile ? localBranchName : ''),
-    };
-    /* eslint-enable @typescript-eslint/naming-convention */
-}
-
-/**
- * `module.rules` shared between the main app build and Storybook (which pushes these onto its own
- * `tools.rspack` config rather than duplicating them, see `.storybook/rsbuild.config.ts`).
- */
-const sharedAssetRules = [
-    // We are importing this worker as a string by using asset/source otherwise it will default to loading via an HTTPS request later.
-    // This causes issues if we have gone offline before the pdfjs web worker is set up as we won't be able to load it from the server.
-    {
-        test: new RegExp('node_modules/pdfjs-dist/legacy/build/pdf.worker.min.mjs'),
-        type: 'asset/source' as const,
-    },
-    {
-        test: /\.lottie$/,
-        type: 'asset/resource' as const,
-    },
-];
-
-/**
- * Get a production grade config for web
- */
-const getCommonConfiguration = ({file = '.env', platform = 'web'}: Environment): Configuration => {
-    const isDevelopment = file === '.env' || file === '.env.development';
-
-    if (!isDevelopment) {
-        const releaseName = `${process.env.npm_package_name}@${process.env.npm_package_version}`;
-        console.debug(`[SENTRY ${platform.toUpperCase()}] Release: ${releaseName}`);
-        console.debug(`[SENTRY ${platform.toUpperCase()}] Assets Path: ${'./dist/**/*.{js,map}'}`);
-    }
-
-    /* eslint-disable @typescript-eslint/naming-convention */
-    return {
-        mode: isDevelopment ? 'development' : 'production',
-        devtool: 'source-map',
-        entry: {
-            main: './index.js',
-        },
-        output: {
-            // Use simple filenames in development to prevent memory leaks from contenthash changes
-            filename: isDevelopment ? '[name].bundle.js' : '[name]-[contenthash].bundle.js',
-            path: path.resolve(dirname, '../../dist'),
-            publicPath: '/',
-            // Wipe the output directory before each build (replaces clean-webpack-plugin)
-            clean: true,
-        },
-        // We can ignore the "module not installed" warning from lottie-react-native
-        // because we are not using the library for JSON format of Lottie animations.
-        ignoreWarnings: [/lottie-react-native\/lib\/module\/LottieView\/index\.web\.js/],
-        // canvaskit-wasm and expo's getBundleUrl.web.ts reference __filename/__dirname, which don't exist in a
-        // browser bundle. Rspack's default ('warn-mock') mocks them to a fixed value but also emits a "Module
-        // parse warning" every time; explicitly opting into 'mock' keeps the same fixed-value behavior without
-        // the warning. This also matters for Storybook's `--smoke-test`, which fails the build on any warning
-        // that isn't explicitly allow-listed.
-        node: {
-            __filename: 'mock',
-            __dirname: 'mock',
-        },
-        plugins: [
-            // Only emit the SW for non-development builds. In dev, the dev-server's HMR
-            // and the SW's caching behavior fight each other and confuse hot reloads.
-            // Remove this guard locally if you want to actually exercise the SW.
-            ...(isDevelopment
-                ? []
-                : [
-                      new GenerateSW({
-                          clientsClaim: true,
-                          skipWaiting: true,
-                          // Cap is generous on purpose: the vendor (~6.5 MiB), main (~5.5 MiB),
-                          // authScreens.prefetch (~6.3 MiB) chunks and canvaskit.wasm (~7.6 MiB) are
-                          // all critical for offline boot, so we precache the lot. Everything in the
-                          // App build is content-hashed, so growth here only costs first-install bytes.
-                          maximumFileSizeToCacheInBytes: 10 * 1024 * 1024,
-                          // Single-page app: any unmatched navigation should serve the cached app shell.
-                          navigateFallback: '/index.html',
-                          // Don't fall back for asset-like or .well-known requests.
-                          navigateFallbackDenylist: [/^\/_/, /^\/\.well-known/, /\/[^/?]+\.[^/]+$/],
-                          runtimeCaching: [
-                              {
-                                  // Same-origin user media (receipts, chat attachments) — cache opportunistically
-                                  // so they're viewable on offline refresh. The function below is serialized into
-                                  // the generated service worker, so it executes in the SW context where
-                                  // `sameOrigin` is the appropriate Workbox match (no need to read `self.location`).
-                                  urlPattern: ({sameOrigin, url, request}) => sameOrigin && (request.destination === 'image' || /\/(receipts|chat-attachments)\//.test(url.pathname)),
-                                  handler: 'StaleWhileRevalidate',
-                                  options: {
-                                      cacheName: 'user-media',
-                                      expiration: {maxEntries: 200, maxAgeSeconds: 7 * 24 * 60 * 60},
-                                  },
-                              },
-                          ],
-                      }),
-                  ]),
-            new rspack.HtmlRspackPlugin({
-                template: 'web/index.html',
-                filename: 'index.html',
-                // Unlike html-webpack-plugin, HtmlRspackPlugin only exposes its own recognized options (title,
-                // filename, meta, etc.) under `htmlRspackPlugin.options` — arbitrary custom keys are silently
-                // unavailable there. Custom values must be passed via `templateParameters` instead, where they
-                // become top-level template variables (e.g. `isStaging`, not `htmlRspackPlugin.options.isStaging`).
-                // A plain object only accepts string values there (real booleans get coerced to the truthy
-                // strings `"true"`/`"false"`), so a function is used to pass real booleans through untouched.
-                templateParameters: (defaultParameters) => ({
-                    ...defaultParameters,
-                    splashLogo: fs.readFileSync(path.resolve(dirname, `../../assets/images/new-expensify${mapEnvironmentToLogoSuffix(file)}.svg`), 'utf-8'),
-                    isWeb: platform === 'web',
-                    isProduction: file === '.env.production',
-                    isStaging: file === '.env.staging',
-                    useThirdPartyScripts: process.env.USE_THIRD_PARTY_SCRIPTS === 'true' || (platform === 'web' && ['.env.production', '.env.staging'].includes(file)),
-                }),
-            }),
-            // Inject <link rel="prefetch" /> into HTML
-            // This is not "webpackPrefetch: true" equivalent!
-            // By convention we use ".prefetch" suffix for such chunks
-            // See RspackPreloadPlugin.ts for why these presets can't use @vue/preload-webpack-plugin directly under Rspack.
-            new RspackPreloadPlugin({
-                rel: 'prefetch',
-                as: 'script',
-                fileWhitelist: [/(.+)\.prefetch(.*)\.js$/],
-            }),
-            new RspackPreloadPlugin({
-                rel: 'preload',
-                as: 'font',
-                fileWhitelist: [/^(?!.*seguiemj).*\.(woff2|ttf)$/],
-            }),
-            new RspackPreloadPlugin({
-                rel: 'prefetch',
-                as: 'fetch',
-                fileWhitelist: [/\.lottie$/],
-            }),
-            new rspack.ProvidePlugin({
-                process: 'process/browser',
-            }),
-
-            // Copies favicons into the dist/ folder to use for unread status
-            new rspack.CopyRspackPlugin({
-                patterns: [
-                    {from: 'web/favicon.png'},
-                    {from: 'web/favicon-unread.png'},
-                    {from: 'web/og-preview-image.png'},
-                    {from: 'web/apple-touch-icon.png'},
-                    {from: 'web/robots.txt'},
-                    {from: 'assets/images/expensify-app-icon.svg'},
-                    {from: 'web/manifest.json'},
-                    {from: 'assets/css', to: 'css'},
-                    {from: 'assets/fonts/web', to: 'fonts'},
-                    {from: 'assets/sounds', to: 'sounds'},
-                    {from: 'assets/pdfs', to: 'pdfs'},
-                    {from: 'node_modules/react-pdf/dist/Page/AnnotationLayer.css', to: 'css/AnnotationLayer.css'},
-                    {from: 'node_modules/react-pdf/dist/Page/TextLayer.css', to: 'css/TextLayer.css'},
-                    {from: '.well-known/apple-app-site-association', to: '.well-known/apple-app-site-association', toType: 'file'},
-                    {from: '.well-known/assetlinks.json', to: '.well-known/assetlinks.json'},
-
-                    // These files are copied over as per instructions here
-                    // https://github.com/wojtekmaj/react-pdf#copying-cmaps
-                    {from: 'node_modules/pdfjs-dist/cmaps/', to: 'cmaps/'},
-
-                    // Group‑IB web SDK injection file
-                    {from: 'web/snippets/gib.js', to: 'gib.js'},
-
-                    // CanvasKit WASM files for @shopify/react-native-skia web support (uses full version)
-                    {from: 'node_modules/canvaskit-wasm/bin/full/canvaskit.wasm'},
-                ],
-            }),
-            new ModuleInitTimingPlugin(),
-            new rspack.EnvironmentPlugin({JEST_WORKER_ID: ''}),
-            new rspack.IgnorePlugin({
-                resourceRegExp: /^\.\/locale$/,
-                contextRegExp: /moment$/,
-            }),
-            ...(file === '.env.production' || file === '.env.staging'
-                ? [
-                      new rspack.IgnorePlugin({
-                          resourceRegExp: /@welldone-software\/why-did-you-render/,
-                      }),
-                  ]
-                : []),
-            ...(platform === 'web' ? [new CustomVersionFilePlugin()] : []),
-            new rspack.DefinePlugin(getDefineValues(file)),
-            ...(isDevelopment ? [] : [new rspack.CssExtractRspackPlugin()]),
-
-            // Upload source maps to Sentry
-            ...(isDevelopment
-                ? []
-                : ([
-                      sentryWebpackPlugin({
-                          authToken: process.env.SENTRY_AUTH_TOKEN as string | undefined,
-                          org: 'expensify',
-                          project: 'app',
-                          release: {
-                              name: `${process.env.npm_package_name}@${process.env.npm_package_version}`,
-                              create: true,
-                              setCommits: {
-                                  auto: true,
-                              },
-                          },
-                          sourcemaps: {
-                              // Use relative path from project root - works for web (dist/)
-                              assets: './dist/**/*.{js,map}',
-                              filesToDeleteAfterUpload: './dist/**/*.map',
-                          },
-                          debug: false,
-                          telemetry: false,
-                      }),
-                  ] as RspackPluginInstance[])),
-
-            // This allows us to interactively inspect JS bundle contents
-            ...(process.env.ANALYZE_BUNDLE === 'true' ? [new BundleAnalyzerPlugin()] : []),
-        ],
-        module: {
-            rules: [
-                {
-                    test: /\.m?js$/,
-                    resolve: {
-                        fullySpecified: false,
-                    },
-                },
-                // Transpiles and lints all the JS
-                {
-                    test: /\.(js|ts)x?$/,
-                    loader: 'babel-loader',
-
-                    /**
-                     * Exclude node_modules except any packages we need to convert for rendering HTML because they import
-                     * "react-native" internally and use JSX which we need to convert to JS for the browser.
-                     *
-                     * You'll need to add anything in here that needs the alias for "react-native" -> "react-native-web"
-                     * You can remove something from this list if it doesn't use "react-native" as an import and it doesn't
-                     * use JSX/JS that needs to be transformed by babel.
-                     */
-                    exclude: [new RegExp(`node_modules/(?!(${includeModules})/).*|\\.native\\.(js|jsx|ts|tsx)$`)],
-                },
-                // Rule for react-native-web-webview
-                {
-                    test: /postMock.html$/,
-                    type: 'asset',
-                    generator: {
-                        filename: '[name].[ext]',
-                    },
-                },
-
-                // Gives the ability to load local images
-                {
-                    test: /\.(png|jpe?g|gif)$/i,
-                    type: 'asset',
-                },
-
-                // Load svg images
-                {
-                    test: /\.svg$/,
-                    resourceQuery: {not: [/raw/]},
-                    exclude: /node_modules/,
-                    use: [
-                        {
-                            loader: '@svgr/webpack',
-                        },
-                    ],
-                },
-                {
-                    test: /\.pdf$/,
-                    type: 'asset',
-                },
-                {
-                    test: /\.css$/i,
-                    use: isDevelopment ? ['style-loader', 'css-loader'] : [rspack.CssExtractRspackPlugin.loader, 'css-loader'],
-                },
-                {
-                    test: /\.(woff2|ttf)$/i,
-                    type: 'asset',
-                },
-                {
-                    resourceQuery: /raw/,
-                    type: 'asset/source',
-                },
-                ...sharedAssetRules,
-                // This prevents import error coming from react-native-tab-view/lib/module/TabView.js
-                // where Pager is imported without extension due to having platform-specific implementations
-                {
-                    test: /\.js$/,
-                    resolve: {
-                        fullySpecified: false,
-                    },
-                    include: [path.resolve(dirname, '../../node_modules/react-native-tab-view/lib/module/TabView.js')],
-                },
-            ],
-        },
-        resolve: {
-            alias: {
-                lodash: 'lodash-es',
-                'react-native-config': 'react-web-config',
-                'react-native$': 'react-native-web',
-                // Use victory-native source files instead of pre-compiled dist (which uses CommonJS exports)
-                'victory-native': path.resolve(dirname, '../../node_modules/victory-native/src/index.ts'),
-                // Required for @shopify/react-native-skia web support
-                'react-native/Libraries/Image/AssetRegistry': false,
-                // Use legacy build of pdfjs-dist to support older browsers
-                'pdfjs-dist$': path.resolve(dirname, '../../node_modules/pdfjs-dist/legacy/build/pdf.mjs'),
-                // Module alias for web
-                // https://webpack.js.org/configuration/resolve/#resolvealias
-                '@assets': path.resolve(dirname, '../../assets'),
-                '@components': path.resolve(dirname, '../../src/components/'),
-                '@hooks': path.resolve(dirname, '../../src/hooks/'),
-                '@libs': path.resolve(dirname, '../../src/libs/'),
-                '@navigation': path.resolve(dirname, '../../src/libs/Navigation/'),
-                '@pages': path.resolve(dirname, '../../src/pages/'),
-                '@prompts': path.resolve(dirname, '../../prompts'),
-                '@styles': path.resolve(dirname, '../../src/styles/'),
-                // This path is provide alias for files like `ONYXKEYS` and `CONST`.
-                '@src': path.resolve(dirname, '../../src/'),
-                '@userActions': path.resolve(dirname, '../../src/libs/actions/'),
-                '@selectors': path.resolve(dirname, '../../src/selectors/'),
-            },
-
-            // Resolve web-specific implementations (`.web.*`) before bare files so React Native
-            // libraries and our own app-level overrides both pick up their browser variants.
-            extensions: ['.web.js', '.js', '.jsx', '.web.ts', '.web.tsx', '.ts', '.tsx'],
-            fallback: {
-                'process/browser': require.resolve('process/browser'),
-                crypto: false,
-                fs: false,
-                path: false,
-            },
-        },
-
-        optimization: {
-            minimizer: [
-                // Preserve the class name for ImageManipulator (expo module), mirroring the TerserPlugin
-                // options used in the webpack config. rspack's types only declare `keep_classnames` as a
-                // boolean, but the underlying SWC minifier accepts a RegExp too (verified: ImageManipulator's
-                // classname survives minification in the output bundle).
-                new rspack.SwcJsMinimizerRspackPlugin({
-                    minimizerOptions: {
-                        compress: {
-                            passes: 2,
-                        },
-                        keep_classnames: /ImageManipulator|ImageModule/,
-                        mangle: {
-                            keep_fnames: true,
-                        },
-                    } as SwcJsMinimizerRspackPluginOptions['minimizerOptions'],
-                }),
-                '...',
-            ],
-            runtimeChunk: 'single',
-            splitChunks: {
-                cacheGroups: {
-                    // We have to load the whole lottie player to get the player to work in offline mode
-                    lottiePlayer: {
-                        test: /[\\/]node_modules[\\/](@dotlottie\/react-player)[\\/]/,
-                        name: 'lottiePlayer',
-                        chunks: 'all',
-                    },
-                    // heic-to library is used sparsely and we want to load it as a separate chunk
-                    // to reduce the potential bundled size of the initial chunk
-                    heicTo: {
-                        test: /[\\/]node_modules[\\/](heic-to)[\\/]/,
-                        name: 'heicTo',
-                        chunks: 'all',
-                        priority: 10, // ensure this chunk has always its own group
-                    },
-                    // ExpensifyIcons chunk - separate chunk loaded eagerly for offline support
-                    expensifyIcons: {
-                        test: /[\\/]src[\\/]components[\\/]Icon[\\/]chunks[\\/]expensify-icons\.chunk\.ts$/,
-                        name: 'expensifyIcons',
-                        chunks: 'all',
-                    },
-                    // Illustrations chunk - separate chunk loaded eagerly for offline support
-                    illustrations: {
-                        test: /[\\/]src[\\/]components[\\/]Icon[\\/]chunks[\\/]illustrations\.chunk\.ts$/,
-                        name: 'illustrations',
-                        chunks: 'all',
-                    },
-                    // Extract all 3rd party dependencies (~75% of App) to separate js file
-                    // This gives a more efficient caching - 3rd party deps don't change as often as main source
-                    // When dependencies don't change rspack would produce the same js file (and content hash)
-                    // After App update end users would download just the main source and resolve the rest from cache
-                    // When dependencies do change cache is invalidated and users download everything - same as before
-                    vendor: {
-                        test: /[\\/]node_modules[\\/]/,
-                        name: 'vendors',
-
-                        // Capture only the scripts needed for the initial load, so any async imports
-                        // would be grouped (and lazy loaded) separately
-                        chunks: 'initial',
-                    },
-                },
-            },
-        },
-    };
-};
-
-/* eslint-enable @typescript-eslint/naming-convention */
-
-export default getCommonConfiguration;
-export {getDefineValues, sharedAssetRules};
