+++ package.json
@@ -31,10 +31,10 @@
     "start": "npx rock start",
     "web": "./scripts/set-pusher-suffix.sh && concurrently npm:web-proxy npm:web-server",
     "web-proxy": "ts-node web/proxy.ts",
-    "web-server": "tsx --expose-gc ./node_modules/.bin/rspack serve --open --config config/rspack/rspack.dev.ts",
-    "build": "tsx ./node_modules/.bin/rspack --config config/rspack/rspack.common.ts --env file=.env.production && tsx ./scripts/combine-web-sourcemaps.ts",
-    "build-staging": "tsx ./node_modules/.bin/rspack --config config/rspack/rspack.common.ts --env file=.env.staging && tsx ./scripts/combine-web-sourcemaps.ts",
-    "build-adhoc": "tsx ./node_modules/.bin/rspack --config config/rspack/rspack.common.ts --env file=.env.adhoc && tsx ./scripts/combine-web-sourcemaps.ts",
+    "web-server": "tsx --expose-gc ./node_modules/.bin/rsbuild dev --open --config config/rsbuild/rsbuild.config.ts",
+    "build": "ENV=production tsx ./node_modules/.bin/rsbuild build --config config/rsbuild/rsbuild.config.ts && tsx ./scripts/combine-web-sourcemaps.ts",
+    "build-staging": "ENV=staging tsx ./node_modules/.bin/rsbuild build --config config/rsbuild/rsbuild.config.ts && tsx ./scripts/combine-web-sourcemaps.ts",
+    "build-adhoc": "ENV=adhoc tsx ./node_modules/.bin/rsbuild build --config config/rsbuild/rsbuild.config.ts && tsx ./scripts/combine-web-sourcemaps.ts",
     "createDocsRoutes": "ts-node .github/scripts/createDocsRoutes.ts",
     "generateAllowedUrls": "ts-node .github/scripts/generateAllowedUrls.ts",
     "detectRedirectCycle": "ts-node .github/scripts/detectRedirectCycle.ts",
@@ -66,12 +66,12 @@
     "storybook-build": "ENV=production storybook build -o dist/docs",
     "storybook-build-staging": "ENV=staging storybook build -o dist/docs",
     "gh-actions-build": "./.github/scripts/buildActions.sh",
-    "analyze-packages": "ANALYZE_BUNDLE=true tsx ./node_modules/.bin/rspack --config config/rspack/rspack.common.ts --env file=.env.production",
+    "analyze-packages": "ANALYZE_BUNDLE=true ENV=production tsx ./node_modules/.bin/rsbuild build --config config/rsbuild/rsbuild.config.ts",
     "symbolicate:android": "npx metro-symbolicate android/app/build/generated/sourcemaps/react/release/index.android.bundle.map",
     "symbolicate:ios": "npx metro-symbolicate main.jsbundle.map",
     "combine-web-sourcemaps": "./scripts/combine-web-sourcemaps.ts",
     "gh-actions-unused-styles": "npx ts-node scripts/findUnusedStyles.ts",
-    "setup-https": "mkcert -install && mkcert -cert-file config/rspack/certificate.pem -key-file config/rspack/key.pem dev.new.expensify.com localhost 127.0.0.1",
+    "setup-https": "mkcert -install && mkcert -cert-file config/rsbuild/certificate.pem -key-file config/rsbuild/key.pem dev.new.expensify.com localhost 127.0.0.1",
     "react-compiler-compliance-check": "ts-node scripts/react-compiler-compliance-check.ts",
     "generate-search-parser": "peggy --format es -o src/libs/SearchParser/searchParser.js src/libs/SearchParser/searchParser.peggy src/libs/SearchParser/baseRules.peggy",
     "generate-autocomplete-parser": "peggy --format es -o src/libs/SearchParser/autocompleteParser.js src/libs/SearchParser/autocompleteParser.peggy src/libs/SearchParser/baseRules.peggy && ./scripts/parser-workletization.sh src/libs/SearchParser/autocompleteParser.js",
@@ -262,9 +262,7 @@
     "@rsbuild/core": "^2.1.4",
     "@rsbuild/plugin-babel": "^2.0.1",
     "@rsbuild/plugin-svgr": "^2.0.5",
-    "@rspack/cli": "^2.1.2",
     "@rspack/core": "^2.1.2",
-    "@rspack/dev-server": "^2.1.0",
     "@rspack/plugin-react-refresh": "^2.0.2",
     "@sentry/cli": "3.3.0",
     "@sentry/webpack-plugin": "5.2.0",
@@ -272,7 +270,6 @@
     "@storybook/addon-docs": "10.4.6",
     "@storybook/cli": "10.4.6",
     "@storybook/react": "10.4.6",
-    "@svgr/webpack": "^6.0.0",
     "@testing-library/react-native": "13.2.0",
     "@types/base-64": "^1.0.2",
     "@types/bun": "^1.3.14",
@@ -303,7 +300,6 @@
     "@vitest/mocker": "4.0.16",
     "@welldone-software/why-did-you-render": "7.0.1",
     "babel-jest": "29.7.0",
-    "babel-loader": "^10.0.0",
     "babel-plugin-module-resolver": "^5.0.0",
     "babel-plugin-react-compiler": "0.0.0-experimental-a1856f3-20260422",
     "babel-plugin-react-native-web": "^0.18.7",
@@ -311,7 +307,6 @@
     "bun": "^1.3.14",
     "concurrently": "^9.2.1",
     "cspell": "9.4.0",
-    "css-loader": "^6.7.2",
     "csv-parse": "^5.5.5",
     "csv-writer": "^1.6.0",
     "diff-so-fancy": "^1.3.0",
@@ -360,7 +355,6 @@
     "source-map": "^0.7.4",
     "storybook": "10.4.6",
     "storybook-react-rsbuild": "^3.3.4",
-    "style-loader": "^3.0.0",
     "svgo": "^4.0.0",
     "ts-jest": "^29.4.9",
     "ts-node": "^10.9.2",
